refactor(bft): extract state, vote_collector, timeouts into submodules
engine.rs was 2166 LOC mixing five concerns:
- timeout constants + propose/prevote/precommit_timeout helpers
- BftPhase enum
- BftRoundState struct + impl (per-(height,round) bookkeeping)
- VoteCollector struct + impl (stake-weighted tally + supermajority)
- the BftEngine itself + tests

Moved the first three into engine/{timeouts,state,vote_collector}.rs
and re-export from engine.rs so the public API at sentrix_bft::engine::*
is unchanged. engine.rs drops to 1944 LOC, focused on BftEngine.

No behaviour change. cargo clippy --workspace --tests -- -D warnings
clean; cargo test --workspace passes; live verified on mainnet 7119
(halt-all + simul-start, all 4 validators advancing 4-of-4 lockstep)
+ testnet 7120 (4 docker validators advancing).

Phase 3 of the workspace audit.
