+++ crates/sentrix-bft/src/engine.rs
@@ -3,8 +3,22 @@
 // Tendermint-style 3-phase: Propose → Prevote → Precommit → Finalize.
 // Proposer selected by weighted round-robin from active DPoS validator set.
 // Finality at 2/3+1 stake weight.
+//
+// State, vote tally, and timeout helpers were extracted to focused
+// submodules (`state.rs`, `vote_collector.rs`, `timeouts.rs`) to keep
+// this file focused on the engine itself.
 
-use serde::{Deserialize, Serialize};
+mod state;
+mod timeouts;
+mod vote_collector;
+
+pub use state::{BftPhase, BftRoundState};
+pub use timeouts::{
+    MAX_ROUND, PRECOMMIT_TIMEOUT_MS, PREVOTE_TIMEOUT_MS, PROPOSE_TIMEOUT_MS, precommit_timeout,
+    prevote_timeout, propose_timeout,
+};
+pub use vote_collector::VoteCollector;
+
 use std::collections::HashMap;
 use std::time::{Duration, Instant};
 // errors used by integration callers, not directly in this module
@@ -12,243 +26,6 @@ use crate::messages::{
     BlockJustification, Precommit, Prevote, RoundStatus, supermajority_threshold,
 };
 use sentrix_staking::StakeRegistry;
-
-// ── Timeouts ─────────────────────────────────────────────────
-
-// Timeouts tuned for 4-validator testnet with ~100ms localhost latency.
-// Round 0 must give enough time for all validators to start up + exchange
-// their first proposal.
-//
-// 2026-04-20 #1d diagnosis (`BFT #1d:` tally logging added in PR #171):
-// the livelock fires when the proposer's request-response Proposal
-// message doesn't reach all peers within the 10s propose timeout.
-// Typically a peer has just reconnected and isn't yet in the proposer's
-// `verified_peers` set at proposal time — the message is silently
-// dropped. The peer's own propose-phase timeout fires with no proposal
-// received, it nil-prevotes, the proposer's supermajority for block
-// fails → nil-precommit chain reaction → skip round, repeat forever.
-// Bumping propose_timeout 10s → 20s widens the window the proposer has
-// to include a freshly-reconnected peer in the outbound send list.
-// It's not a root-cause fix (that needs proposer re-broadcast or
-// verified-peer stability, backlog #1d follow-up) but empirically
-// stops the stall. prevote/precommit also nudged up to 12s so phase
-// transitions don't race the wider propose window.
-pub const PROPOSE_TIMEOUT_MS: u64 = 20_000;
-pub const PREVOTE_TIMEOUT_MS: u64 = 12_000;
-pub const PRECOMMIT_TIMEOUT_MS: u64 = 12_000;
-pub const TIMEOUT_INCREMENT_MS: u64 = 1_000; // +1s per round for propose
-pub const VOTE_TIMEOUT_INCREMENT_MS: u64 = 2_000; // +2s per round for votes
-pub const MAX_TIMEOUT_MS: u64 = 30_000;
-pub const MAX_ROUND: u32 = 100;
-
-// ── BFT State ────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum BftPhase {
-    Propose,
-    Prevote,
-    Precommit,
-    Finalize,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BftRoundState {
-    pub height: u64,
-    pub round: u32,
-    pub phase: BftPhase,
-    /// Block hash proposed in this round (if any)
-    pub proposed_hash: Option<String>,
-    /// Prevotes collected: validator → (block_hash option, stake_weight)
-    pub prevotes: HashMap<String, (Option<String>, u64)>,
-    /// Precommits collected: validator → (block_hash option, signature bytes, stake_weight).
-    /// V1 (re-applied 2026-04-25 on v2.1.16 base): signature bytes kept
-    /// alongside vote so `BlockJustification` emits REAL signatures at
-    /// finalize, not `vec![]` placeholders. Closes V1 Voyager blocker.
-    ///
-    /// Prior v2.1.14 revert was the wrong diagnosis — the real v2.1.12
-    /// livelock trigger was PR #244's hot-path fsync (closed v2.1.15).
-    /// With V2 M-15 locked-block re-propose now shipped in v2.1.16 too,
-    /// the locked-validator-prevote-nil scenario self-unsticks via
-    /// re-propose instead of livelocking.
-    pub precommits: HashMap<String, (Option<String>, Vec<u8>, u64)>,
-    /// Our own vote cast this round (prevent double-voting)
-    pub our_prevote_cast: bool,
-    pub our_precommit_cast: bool,
-    /// Locked value: if we precommitted for a hash, we're locked on it
-    pub locked_hash: Option<String>,
-    pub locked_round: Option<u32>,
-    /// V2 M-15 Step 1: block bytes cached alongside `locked_hash` so a
-    /// locked validator elected proposer in a later round can RE-PROPOSE
-    /// the cached block instead of building a fresh one (whose timestamp
-    /// → different hash would trigger the lock-nil-prevote livelock).
-    /// See `audits/v2-locked-block-repropose-implementation-plan.md`.
-    ///
-    /// Populated from `staging_block` at the moment prevote-supermajority
-    /// fires. Cleared on `new_height`. Preserved across `advance_round`
-    /// (same semantics as `locked_hash`). `#[serde(default)]` for
-    /// forward-compat — `BftRoundState` is in-memory-only today but the
-    /// planned WAL (M-16 follow-up) will persist this field.
-    #[serde(default)]
-    pub locked_block: Option<Vec<u8>>,
-    /// V2 M-15 Step 1: temporary hold for a proposal's block bytes
-    /// between `on_proposal` arrival and prevote-supermajority. If the
-    /// staged hash matches the hash that crosses 2/3+ prevote threshold,
-    /// bytes are promoted to `locked_block`. Cleared on `advance_round`
-    /// and `new_height`. Not wired yet — Step 2 adds the on_proposal /
-    /// on_own_proposal parameters to stash bytes here.
-    #[serde(default)]
-    pub staging_block: Option<(String, Vec<u8>)>,
-    /// Total stake of current active set (for threshold calculation)
-    pub total_active_stake: u64,
-}
-
-impl BftRoundState {
-    pub fn new(height: u64, round: u32, total_active_stake: u64) -> Self {
-        Self {
-            height,
-            round,
-            phase: BftPhase::Propose,
-            proposed_hash: None,
-            prevotes: HashMap::new(),
-            precommits: HashMap::new(),
-            our_prevote_cast: false,
-            our_precommit_cast: false,
-            locked_hash: None,
-            locked_round: None,
-            locked_block: None,
-            staging_block: None,
-            total_active_stake,
-        }
-    }
-
-    pub fn advance_round(&mut self) {
-        self.round = self.round.saturating_add(1);
-        self.phase = BftPhase::Propose;
-        self.proposed_hash = None;
-        self.prevotes.clear();
-        self.precommits.clear();
-        self.our_prevote_cast = false;
-        self.our_precommit_cast = false;
-        // locked_hash, locked_round, and locked_block persist across rounds
-        // (same PoLC semantics). staging_block is per-round and cleared here —
-        // next round's proposal will re-stash.
-        self.staging_block = None;
-    }
-
-    pub fn advance_height(&mut self, new_height: u64, total_active_stake: u64) {
-        *self = Self::new(new_height, 0, total_active_stake);
-    }
-}
-
-// ── Vote Collector ───────────────────────────────────────────
-
-#[derive(Debug, Clone, Default)]
-pub struct VoteCollector {
-    /// Track which block_hash has how much stake in prevotes
-    prevote_tally: HashMap<Option<String>, u64>,
-    /// Same for precommits
-    precommit_tally: HashMap<Option<String>, u64>,
-}
-
-impl VoteCollector {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn add_prevote(&mut self, block_hash: Option<String>, stake: u64) {
-        *self.prevote_tally.entry(block_hash).or_insert(0) += stake;
-    }
-
-    pub fn add_precommit(&mut self, block_hash: Option<String>, stake: u64) {
-        *self.precommit_tally.entry(block_hash).or_insert(0) += stake;
-    }
-
-    /// Check if any single block_hash has supermajority prevotes
-    pub fn prevote_supermajority(&self, total_stake: u64) -> Option<Option<String>> {
-        let threshold = supermajority_threshold(total_stake);
-        for (hash, &weight) in &self.prevote_tally {
-            if weight >= threshold {
-                return Some(hash.clone());
-            }
-        }
-        None
-    }
-
-    /// Check if any single block_hash has supermajority precommits
-    pub fn precommit_supermajority(&self, total_stake: u64) -> Option<Option<String>> {
-        let threshold = supermajority_threshold(total_stake);
-        for (hash, &weight) in &self.precommit_tally {
-            if weight >= threshold {
-                return Some(hash.clone());
-            }
-        }
-        None
-    }
-
-    /// Total prevote weight collected
-    pub fn total_prevote_weight(&self) -> u64 {
-        self.prevote_tally.values().sum()
-    }
-
-    /// Total precommit weight collected
-    pub fn total_precommit_weight(&self) -> u64 {
-        self.precommit_tally.values().sum()
-    }
-
-    /// Snapshot of the current precommit tally, for diagnostic logging.
-    /// Returns `(short_hash_or_nil, weight)` pairs sorted by weight desc
-    /// so split-vote livelocks are easy to spot in journalctl.
-    ///
-    /// Added for backlog #1d — the nil-skip branch previously lost this
-    /// information, making unanimous-nil-timeout indistinguishable from
-    /// split-vote livelock in the logs.
-    pub fn precommit_tally_snapshot(&self) -> Vec<(String, u64)> {
-        let mut entries: Vec<(String, u64)> = self
-            .precommit_tally
-            .iter()
-            .map(|(h, w)| {
-                let label = match h {
-                    Some(hash) => {
-                        let trimmed = &hash[..hash.len().min(12)];
-                        format!("{trimmed}…")
-                    }
-                    None => "nil".to_string(),
-                };
-                (label, *w)
-            })
-            .collect();
-        entries.sort_by_key(|e| std::cmp::Reverse(e.1));
-        entries
-    }
-
-    pub fn reset(&mut self) {
-        self.prevote_tally.clear();
-        self.precommit_tally.clear();
-    }
-}
-
-// ── Timeout Calculator ───────────────────────────────────────
-
-pub fn propose_timeout(round: u32) -> Duration {
-    let ms = PROPOSE_TIMEOUT_MS
-        .saturating_add((round as u64).saturating_mul(TIMEOUT_INCREMENT_MS))
-        .min(MAX_TIMEOUT_MS);
-    Duration::from_millis(ms)
-}
-
-pub fn prevote_timeout(round: u32) -> Duration {
-    let ms = PREVOTE_TIMEOUT_MS
-        .saturating_add((round as u64).saturating_mul(VOTE_TIMEOUT_INCREMENT_MS))
-        .min(MAX_TIMEOUT_MS);
-    Duration::from_millis(ms)
-}
-
-pub fn precommit_timeout(round: u32) -> Duration {
-    let ms = PRECOMMIT_TIMEOUT_MS
-        .saturating_add((round as u64).saturating_mul(VOTE_TIMEOUT_INCREMENT_MS))
-        .min(MAX_TIMEOUT_MS);
-    Duration::from_millis(ms)
-}
 
 // ── BFT Engine ───────────────────────────────────────────────
